requirements.txt
# ============================================================
# Feature Store Requirements
# Compatible versions for Feast, PySpark, Kafka, and Airflow
# ============================================================

# Core Feature Store
feast[redis,postgres]==0.37.1

# Apache Spark
pyspark==3.5.0

# Kafka Integration
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data Processing
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<2.0.0
pyarrow>=14.0.0,<16.0.0

# Database Connectors
psycopg2-binary==2.9.9
redis>=4.5.0,<5.0.0
sqlalchemy>=2.0.0,<2.1.0

# Apache Airflow (with compatible constraints)
apache-airflow==2.8.1
apache-airflow-providers-apache-kafka==1.3.1
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-redis==3.6.0

# Delta Lake
delta-spark==3.0.0

# ML Libraries
scikit-learn>=1.3.0,<1.4.0
xgboost>=2.0.0,<2.1.0

# Data Validation & Quality
great-expectations>=0.18.0,<0.19.0
pandera>=0.17.0,<0.18.0

# Notebook & Visualization
jupyterlab>=4.0.0,<4.1.0
matplotlib>=3.7.0,<3.9.0
seaborn>=0.12.0,<0.14.0
plotly>=5.18.0,<5.19.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
requests>=2.31.0
httpx>=0.25.0

# Type Checking & Linting
mypy>=1.5.0
black>=23.9.0
ruff>=0.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
